finally got tests working, but lution is wrong

+++ main.py
@@ -167,7 +167,7 @@ def array_and_plot_grid(input_grid):
 
 
 def grid2rgb(input_grid):
-    grid = np.array(input_grid)
+    grid = np.array(input_grid).astype(np.int64)
     rgb_grid = rgb_lookup()[grid]
     return rgb_grid
